+++ packages/react-form-state/CHANGELOG.md
@@ -7,4 +7,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Changed
+
 - Validators based on `validate` now always succeed for an empty input. The `required` and `requiredString` validators continue to behave the same way they used to.
+- The `FieldDescriptors` type is now named `Fields`.